pyproject.toml
[tool.poetry]
name = "heritai"
version = "0.1.0"
description = "heritage application POC"
authors = ["carlosprietofernandez-tomtom <carlos.prietofernandez@tomtom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.25.0"
plotly = "^5.16.1"
streamlit-authenticator = "^0.2.2"
langchain = "^0.0.271"
openai = "^0.27.9"
chromadb = "^0.4.6"
tiktoken = "^0.4.0"
streamlit-option-menu = "^0.3.6"
pypdf2 = "^3.0.1"
faiss-cpu = "^1.7.4"
pypdf = "^3.15.2"
graphviz = "^0.20.1"
pip = "^23.2.1"
streamlit-analytics = "^0.4.1"
python-dotenv = "^1.0.0"
nbformat = "^5.9.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
